+++ lib/src/src.dart
@@ -1,3 +1,9 @@
 export 'boxes/boxes.dart'
-    show HivezBox, HivezBoxLazy, HivezBoxIsolated, HivezBoxIsolatedLazy;
+    show
+        HivezBox,
+        HivezBoxLazy,
+        HivezBoxIsolated,
+        HivezBoxIsolatedLazy,
+        BaseHivezBox,
+        BoxInterface;
 export 'extensions/extensions.dart';