+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Added `moveKey` method to reassign a value from one key to another (renames the key while preserving the value).
 - Added `foreachKey` and `foreachValue` methods to iterate over all keys and values in the box
+- Made the base `BoxInterface` class simpler for better abstraction and flexibility
 
 ## 0.0.9